refactor(geral): organizando pastas e criando documentação das migrations
