+++ frontend/src/App.tsx
@@ -1,24 +1,24 @@
 import { Route, Routes } from 'react-router-dom';
-import NotFound from './pages/NotFound';
-import Login from './pages/Login';
+import AdminPage from './pages/Admin/AdminPage';
+import CreateCourse from './pages/Admin/CreateCourse';
+import EditCourse from './pages/Admin/EditCourse';
+import Students from './pages/Admin/Students';
+import BookmarkedCourses from './pages/Course/BookmarkedCourses';
+import ConfirmEmail from './pages/Course/ConfirmEmail';
+import CourseModules from './pages/Course/CourseModules';
+import CreateAccount from './pages/Course/CreateAccount';
+import ForgotPassword from './pages/Course/ResetPassword';
 import Header from './components/Header';
+import Homepage from './pages/Course/Homepage';
+import LessonPage from './pages/Course/LessonPage';
+import Lessons from './pages/Course/Lessons';
+import Login from './pages/Course/Login';
+import NotFound from './pages/Course/NotFound';
+import PdfPage from './pages/Course/PdfPage';
+import Profile from './pages/Course/Profile';
+import StudentCourses from './pages/Course/Courses';
+import SupportPage from './pages/Course/SupportPage';
 import Footer from './components/Footer';
-import CreateAccount from './pages/CreateAccount';
-import StudentCourses from './pages/Courses';
-import ConfirmEmail from './pages/ConfirmEmail';
-import ForgotPassword from './pages/ResetPassword';
-import Homepage from './pages/Homepage';
-import AdminPage from './pages/AdminPage';
-import CourseModules from './pages/CourseModules';
-import Lessons from './pages/Lessons';
-import BookmarkedCourses from './pages/BookmarkedCourses';
-import Profile from './pages/Profile';
-import LessonPage from './pages/LessonPage';
-import CreateCourse from './pages/CreateCourse';
-import EditCourse from './pages/EditCourse';
-import SupportPage from './pages/SupportPage';
-import PdfPage from './pages/PdfPage';
-import Students from './pages/Students';
 import './App.css';
 
 function App() {
@@ -27,10 +27,14 @@ function App() {
       <Header />
       <Routes>
         <Route path="/" element={ <Homepage /> } />
-        <Route path="/login" element={ <Login /> } />
-        <Route path="/create-account" element={ <CreateAccount /> } />
-        <Route path="/courses" element={ <StudentCourses /> } />
+        <Route path="/admin" element={ <AdminPage /> } />
+        <Route path="/admin/courses" element="Cursos" />
+        <Route path="/admin/create" element={ <CreateCourse /> } />
+        <Route path="/admin/edit" element={ <EditCourse /> } />
+        <Route path="/admin/students" element={ <Students /> } />
         <Route path="/bookmarked" element={ <BookmarkedCourses /> } />
+        <Route path="/confirm/:token" element={ <ConfirmEmail /> } />
+        <Route path="/courses" element={ <StudentCourses /> } />
         <Route path="/courses/:courseId/modules" element={ <CourseModules /> } />
         <Route
           path="/courses/:courseId/modules/:moduleId/lessons"
@@ -40,19 +44,15 @@ function App() {
           path="/courses/:courseId/modules/:moduleId/lessons/:lessonId"
           element={ <LessonPage /> }
         />
-        <Route path="/profile" element={ <Profile /> } />
-        <Route path="/confirm/:token" element={ <ConfirmEmail /> } />
-        <Route path="/reset-password/:token" element={ <ForgotPassword /> } />
-        <Route path="/admin" element={ <AdminPage /> } />
-        <Route path="/admin/create" element={ <CreateCourse /> } />
-        <Route path="/admin/edit" element={ <EditCourse /> } />
-        <Route path="/admin/students" element={ <Students /> } />
-        <Route path="/admin/courses" element="Cursos" />
-        <Route path="/support" element={ <SupportPage /> } />
         <Route
           path="/courses/:courseId/modules/:moduleId/lessons/:lessonId/pdfs"
           element={ <PdfPage /> }
         />
+        <Route path="/create-account" element={ <CreateAccount /> } />
+        <Route path="/login" element={ <Login /> } />
+        <Route path="/profile" element={ <Profile /> } />
+        <Route path="/reset-password/:token" element={ <ForgotPassword /> } />
+        <Route path="/support" element={ <SupportPage /> } />
         <Route path="*" element={ <NotFound /> } />
       </Routes>
       <Footer />
