+++ frontend/src/pages/Course/BookmarkedCourses.tsx
@@ -1,9 +1,9 @@
 import { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import CoursesBackground from '../components/CoursesBackground';
-import { requestData, requestUpdate, setToken } from '../services/requests';
-import { UserCourses } from '../types/courseType';
-import CourseCard from '../components/CourseCard';
+import CoursesBackground from '../../components/CoursesBackground';
+import { requestData, requestUpdate, setToken } from '../../services/requests';
+import { UserCourses } from '../../types/courseType';
+import CourseCard from '../../components/CourseCard';
 
 function BookmarkedCourses() {
   const [courses, setCourses] = useState<UserCourses[]>([]);
