+++ frontend/src/pages/Course/Courses.tsx
@@ -1,9 +1,9 @@
 import { useState, useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
-import CoursesBackground from '../components/CoursesBackground';
-import { requestData, requestUpdate, setToken } from '../services/requests';
-import { UserCourses } from '../types/courseType';
-import CourseCard from '../components/CourseCard';
+import CoursesBackground from '../../components/CoursesBackground';
+import { requestData, requestUpdate, setToken } from '../../services/requests';
+import { UserCourses } from '../../types/courseType';
+import CourseCard from '../../components/CourseCard';
 
 function StudentCourses() {
   const [courses, setCourses] = useState<UserCourses[]>([]);
