+++ frontend/src/pages/Course/NotFound.tsx
