+++ frontend/src/pages/Course/Profile.tsx
@@ -1,12 +1,12 @@
 import { Input, Typography } from '@material-tailwind/react';
 import { useEffect, useState, MouseEvent } from 'react';
 import { useNavigate } from 'react-router-dom';
-import CoursesBackground from '../components/CoursesBackground';
-import OrangeButton from '../components/OrangeButton';
-import EyeButton from '../components/EyeButton';
-import { requestPost, requestUpdate, setToken } from '../services/requests';
-import WarnigIcon from '../components/WarningIcon';
-import { UserType, initialUserState } from '../types/userTypes';
+import CoursesBackground from '../../components/CoursesBackground';
+import OrangeButton from '../../components/OrangeButton';
+import EyeButton from '../../components/EyeButton';
+import { requestPost, requestUpdate, setToken } from '../../services/requests';
+import WarnigIcon from '../../components/WarningIcon';
+import { UserType, initialUserState } from '../../types/userTypes';
 
 function Profile() {
   const [user, setUser] = useState<UserType>(initialUserState);
